Move custom guard classes for testing purposes into their respective test file
