+++ packages/meteor-type-validation/tests/type-tests/guards/CustomGuards.test-d.ts
@@ -1,11 +1,30 @@
-import { defineMethods, definePublications, UserLoggedInGuard } from '@meteor-type-validation/server';
+import { defineMethods, definePublications, Guard, UserLoggedInGuard } from '@meteor-type-validation/server';
+import { Meteor } from 'meteor/meteor';
+import * as v from 'valibot';
 import { describe, expectTypeOf, it } from 'vitest';
-import { AdminGuard } from '../../lib/AdminGuard';
-import { CreatedByCurrentUser } from '../../lib/CreatedByCurrentUserGuard';
-import { QueryValidationGuard } from '../../lib/QueryValidationGuard';
 import { EditTodoSchema } from '../../lib/Schemas';
 
 describe('CreatedByCurrentUser', () => {
+    class CreatedByCurrentUser extends Guard {
+        public readonly writeToParams = false;
+        public readonly writeToContext = false;
+        
+        public readonly contextSchema = UserLoggedInGuard.contextSchema;
+        
+        public readonly paramSchema = [
+            v.object({
+                createdBy: v.string(),
+            }),
+        ]
+        
+        public validate() {
+            this.assertContext();
+            if (this.params[0].createdBy !== this.context.userId) {
+                throw new Meteor.Error(401, 'You do not have permission for this resource');
+            }
+        }
+    }
+    
     describe('methods', () => {
         it(`should infer that the user is logged in from the 'this' context`, () => {
             defineMethods({
@@ -60,6 +79,22 @@ describe('CreatedByCurrentUser', () => {
 })
 
 describe('AdminGuard', () => {
+    class AdminGuard extends Guard {
+        
+        public readonly writeToParams = false;
+        public readonly paramSchema = [];
+        public readonly writeToContext = false;
+        
+        public readonly contextSchema = v.objectAsync({
+            userId: v.string(),
+            user: v.pipe(v.any(), v.object({
+                roles: v.pipe(
+                    v.array(v.picklist(['admin'])),
+                )
+            })),
+        });
+    }
+    
     describe('methods', () => {
         it(`should assert that the user's 'roles' field includes 'admin'`, () => {
             defineMethods({
@@ -124,6 +159,36 @@ describe('PermissionGuard', () => {
 })
 
 describe('QueryValidationGuard', () => {
+    class QueryValidationGuard extends Guard {
+        
+        public readonly writeToParams = false;
+        public readonly writeToContext = false;
+        public readonly contextSchema = v.object({})
+        public readonly paramSchema = [
+            v.object({
+                channelId: v.string(),
+            }),
+            v.object({
+                fields: v.record(
+                    v.picklist([
+                        'title',
+                        'message',
+                        'user.name',
+                        'createdAt'
+                    ]),
+                    v.literal(1)
+                ),
+                limit: v.pipe(
+                    v.number(),
+                    v.integer(),
+                    v.minValue(1),
+                    v.maxValue(100)
+                ),
+            })
+        ];
+        
+    }
+    
     describe('methods', () => {
         it(`can infer parameter types only from the guard's schema type`, () => {
             defineMethods({
